Full upload
This is the full wokring compiler for the PL/0 language. It split into
four main components: scanner, parser, virtual machine, and error
recovery. The scanner takes the high level language and outputs tokens,
these tokens are then ran through the parser to check for errors and
outputs assembly language. The assembly language is then feed into the
machine dependent virtual machine, which outputs some result. Error
recovery is implemented last in order to make debugger easier for the
coder, and allows for non-serious coding errors to unhinder the
execution process.